feat: rename accounts to core
