+++ src/core/users/SignOut/SignOut.styled.ts
